requirements.txt
opencv-python>=4.9.0
pytesseract>=0.3.10
Pillow>=10.3.0
numpy>=1.26.0
torch
torchvision
diffusers
transformers
accelerate
